Switch content script back to declarative injection.
I could not find a good event to fire rogrammatic injection, so I decided to switch it back.
But in this commit, content script will not send message to background to ask distribute profile,
I create a current profile cache in chrome.storage, content script can read and watch it directly.

Signed-off-by: Sunny <sunny@magicgestures.org>